+++ kowalski/Cargo.toml
@@ -24,11 +24,11 @@ keywords = [
 [dependencies]
 # Core components are non-optional
 kowalski-core = { path = "../kowalski-core", version = "1.0.0" }
-
-# Optional CLI
-kowalski-cli = { path = "../kowalski-cli", version = "1.0.0", optional = true }
+kowalski-cli = { path = "../kowalski-cli", version = "1.0.0" }
 
 tokio = { workspace = true }
+clap = { version = "4.5", features = ["derive"] }
+env_logger = { workspace = true }
 
 [lib]
 name = "kowalski"
@@ -38,7 +38,7 @@ path = "src/lib.rs"
 default = []
 
 # Other features
-cli = ["dep:kowalski-cli"]
+cli = []
 postgres = ["kowalski-core/postgres"]
 
 # All features (CLI + PostgreSQL client stack)
